undo synchronized()
The onReceive() method of a BroadcastReceiver is always called on the application's main thread. The Android system ensures that intents are delivered sequentially to the receiver. This means there is no risk of concurrent execution of the onReceive() method for a single BroadcastReceiver instance, and therefore no need for synchronization to protect its instance variables.

https://github.com/xseignard/cordovarduino/commit/51f994137569d75b788366f645e4ca3dd3e7ea9e#diff-1207af4efda49c423fb6c7981b72b44d4bd1e97f488d5590b8deb94a62560928R47-R48